+++ client/packages/host/src/routers/ProgramRouter.tsx
@@ -2,8 +2,8 @@ import React, { FC } from 'react';
 import { RouteBuilder, Navigate, useMatch } from '@openmsupply-client/common';
 import { AppRoute } from '@openmsupply-client/config';
 
-const ImmunisationProgramService = React.lazy(
-  () => import('@openmsupply-client/system/src/Immunisation/Service/Service')
+const ProgramService = React.lazy(
+  () => import('@openmsupply-client/programs/src/Service/Service')
 );
 
 const immunisationFullPath = RouteBuilder.create(AppRoute.Programs)
@@ -15,7 +15,7 @@ export const ProgramsRouter: FC = () => {
   const gotoImmunisations = useMatch(immunisationFullPath);
 
   if (gotoImmunisations) {
-    return <ImmunisationProgramService />;
+    return <ProgramService />;
   }
 
   const notFoundRoute = RouteBuilder.create(AppRoute.PageNotFound).build();
