+++ client/packages/programs/src/index.ts
@@ -4,3 +4,5 @@ export * from './hooks';
 export * from './JsonForms';
 export * from './schema_types';
 export * from './utils';
+export * from './VaccineCourseEditModal';
+export * from './Service';
