--- client/packages/system/src/Immunisation/api/hooks/index.ts
@@ -1,6 +1,0 @@
-export * from './keys';
-export * from './useImmunisationProgram';
-export * from './useImmunisationProgramList';
-export * from './useVaccineCourse';
-export * from './useVaccineCourseList';
-export * from './types';
