--- client/packages/system/src/Immunisation/api/index.ts
@@ -1,2 +1,0 @@
-export * from './hooks';
-export * from './operations.generated';
