--- client/packages/system/src/Immunisation/api/operations.graphql
@@ -1,207 +1,0 @@
-fragment ImmunisationProgram on ProgramNode {
-  id
-  name
-  vaccineCourses {
-    name
-  }
-}
-
-fragment VaccineCourseSchedule on VaccineCourseScheduleNode {
-  id
-  doseNumber
-  label
-}
-
-fragment VaccineCourseItem on VaccineCourseItemNode {
-  id
-  itemId
-  name
-}
-
-query programs(
-  $storeId: String
-  $first: Int
-  $offset: Int
-  $key: ProgramSortFieldInput!
-  $desc: Boolean
-  $filter: ProgramFilterInput
-) {
-  programs(
-    storeId: $storeId
-    page: { first: $first, offset: $offset }
-    sort: { key: $key, desc: $desc }
-    filter: $filter
-  ) {
-    ... on ProgramConnector {
-      __typename
-      nodes {
-        __typename
-        ...ImmunisationProgram
-      }
-      totalCount
-    }
-  }
-}
-
-mutation insertImmunisationProgram(
-  $input: InsertImmunisationProgramInput
-  $storeId: String
-) {
-  centralServer {
-    program {
-      insertImmunisationProgram(input: $input, storeId: $storeId) {
-        __typename
-        ... on ProgramNode {
-          ...ImmunisationProgram
-        }
-        ... on InsertImmunisationProgramError {
-          __typename
-          error {
-            __typename
-            description
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation updateImmunisationProgram(
-  $input: UpdateImmunisationProgramInput
-  $storeId: String
-) {
-  centralServer {
-    program {
-      updateImmunisationProgram(input: $input, storeId: $storeId) {
-        __typename
-        ... on ProgramNode {
-          ...ImmunisationProgram
-        }
-        ... on UpdateImmunisationProgramError {
-          __typename
-          error {
-            ... on UniqueValueViolation {
-              __typename
-              field
-            }
-            description
-          }
-        }
-      }
-    }
-  }
-}
-fragment VaccineCourse on VaccineCourseNode {
-  id
-  name
-  programId
-  demographicIndicatorId
-  doses
-  coverageRate
-  wastageRate
-  isActive
-  demographicIndicator {
-    name
-    id
-    baseYear
-  }
-  vaccineCourseItems {
-    ...VaccineCourseItem
-  }
-  vaccineCourseSchedules {
-    ...VaccineCourseSchedule
-  }
-}
-
-query vaccineCourses(
-  $first: Int
-  $offset: Int
-  $key: VaccineCourseSortFieldInput!
-  $desc: Boolean
-  $filter: VaccineCourseFilterInput
-) {
-  vaccineCourses(
-    page: { first: $first, offset: $offset }
-    sort: { key: $key, desc: $desc }
-    filter: $filter
-  ) {
-    ... on VaccineCourseConnector {
-      __typename
-      nodes {
-        ...VaccineCourse
-      }
-      totalCount
-    }
-  }
-}
-
-mutation insertVaccineCourse(
-  $input: InsertVaccineCourseInput
-  $storeId: String
-) {
-  centralServer {
-    vaccineCourse {
-      insertVaccineCourse(input: $input, storeId: $storeId) {
-        __typename
-        ... on VaccineCourseNode {
-          ...VaccineCourse
-        }
-        ... on InsertVaccineCourseError {
-          __typename
-          error {
-            description
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation updateVaccineCourse(
-  $input: UpdateVaccineCourseInput
-  $storeId: String
-) {
-  centralServer {
-    vaccineCourse {
-      updateVaccineCourse(input: $input, storeId: $storeId) {
-        __typename
-        ... on VaccineCourseNode {
-          __typename
-          ...VaccineCourse
-        }
-        ... on UpdateVaccineCourseError {
-          __typename
-          error {
-            description
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation deleteImmunisationProgram($immunisationProgramId: String!) {
-  centralServer {
-    program {
-      deleteImmunisationProgram(immunisationProgramId: $immunisationProgramId) {
-        ... on DeleteResponse {
-          __typename
-          id
-        }
-      }
-    }
-  }
-}
-
-mutation deleteVaccineCourse($vaccineCourseId: String!) {
-  centralServer {
-    vaccineCourse {
-      deleteVaccineCourse(vaccineCourseId: $vaccineCourseId) {
-        ... on DeleteResponse {
-          __typename
-          id
-        }
-      }
-    }
-  }
-}
