--- client/packages/system/src/Immunisation/index.ts
@@ -1,3 +1,0 @@
-export * from './Service';
-export * from './api';
-export * from './VaccineCourseEditModal';
